+++ src/date-input.component.jsx
@@ -11,6 +11,7 @@ import 'react-day-picker/lib/style.css';
 // App imports
 import TimePicker from './time-picker/time-picker.component';
 import YearMonthPicker from './year-month-picker/year-month-picker.component';
+import Navbar from './navbar/navbar.component';
 import './date-input.scss';
 
 // Date formats used by the component (mainly by the getDate method)
@@ -20,41 +21,59 @@ const FORMATS = {
   DATE_OBJECT: 'DATE_OBJECT',
 };
 
+const classPrefix = 'oc-datetime';
+
 export default class DateInput extends React.Component {
   static propTypes = {
+    className: PropTypes.string,
     value: PropTypes.string,
     onChange: PropTypes.func,
+    onDayClick: PropTypes.func,
     locale: PropTypes.string,
     dateFormat: PropTypes.string,
     inputProps: PropTypes.object,
     inputRef: PropTypes.func,
     disabled: PropTypes.bool,
+    selectedDays: PropTypes.oneOfType([
+      PropTypes.object,
+      PropTypes.func,
+      PropTypes.array,
+    ]),
+    showOverlay: PropTypes.bool,
     showWeekNumbers: PropTypes.bool,
+    showClearValue: PropTypes.bool,
     time: PropTypes.bool,
     minutesInterval: PropTypes.number,
   };
 
   static defaultProps = {
+    className: '',
     value: '',
     dateFormat: 'L',
-    locale: 'en',
+    locale: 'en-GB',
     onChange() {
+    },
+    onDayClick: () => {
     },
     inputProps: {},
     inputRef() {
     },
     disabled: false,
+    selectedDays: null,
+    showOverlay: false,
     showWeekNumbers: true,
+    showClearValue: true,
     time: false,
     minutesInterval: 5,
   };
 
   static getDerivedStateFromProps(props, state) {
-    if (props.value !== state.lastValue) {
+    if (!state.showOverlay && props.value !== state.lastValue) {
       const momentDate = moment.utc(props.value, moment.ISO_8601);
       return {
         lastValue: props.value,
         selectedDay: DateInput.getDate(momentDate, FORMATS.DATE_OBJECT),
+        showOverlay: props.showOverlay || state.showOverlay,
         inputDate: DateInput.getDate(momentDate, FORMATS.PRETTY_DATE, props.dateFormat),
       };
     }
@@ -102,12 +121,18 @@ export default class DateInput extends React.Component {
 
     this.localeUtils = Object.assign(
       LocaleUtils,
-      { getFirstDayOfWeek: () => moment.localeData().firstDayOfWeek() },
+      {
+        getFirstDayOfWeek: () => moment.localeData()
+          .firstDayOfWeek(),
+      },
     );
 
     this.input = null;
     this.dayPicker = null;
-    this.focused = false;
+
+    // Used in onBlur handler to determine whether or not blur happened because of a click
+    // on the overlay
+    this.mouseClickedOnContainer = false;
   }
 
   componentWillUnmount() {
@@ -122,9 +147,9 @@ export default class DateInput extends React.Component {
     if (!this.calendarContainer) return;
 
     // Closes overlay if user clicks outside the calendar (and input field)
-    if (!this.calendarContainer.contains(e.target) &&
-      this.state.showOverlay &&
-      e.target !== this.input) {
+    if (!this.calendarContainer.contains(e.target)
+      && this.state.showOverlay
+      && e.target !== this.input) {
       this.closeOverlay();
       document.removeEventListener('click', this.onDocumentClick);
     }
@@ -134,7 +159,8 @@ export default class DateInput extends React.Component {
    * Returns the first of the week based on locale (used by DayPicker)
    * @returns {number}
    */
-  getFirstDayOfWeek = () => moment.localeData(this.props.locale).firstDayOfWeek();
+  getFirstDayOfWeek = () => moment.localeData(this.props.locale)
+    .firstDayOfWeek();
 
   /**
    * Handles input focus event. Shows an overlay and adds an click event listener to the document
@@ -142,7 +168,6 @@ export default class DateInput extends React.Component {
    */
   handleInputFocus = (e) => {
     const { showOverlay, selectedDay } = this.state;
-    this.focused = true;
 
     this.setState({
       showOverlay: true,
@@ -162,8 +187,6 @@ export default class DateInput extends React.Component {
    * @param e
    */
   closeOverlay = (e) => {
-    this.focused = false;
-
     this.setState({
       showOverlay: false,
     }, () => {
@@ -182,7 +205,8 @@ export default class DateInput extends React.Component {
 
     this.setState({ inputDate });
     // This fires only if the new date is valid in given format
-    if (moment.utc(inputDate, dateFormat).isValid() && this.isValidFormat(inputDate)) {
+    if (moment.utc(inputDate, dateFormat)
+      .isValid() && this.isValidFormat(inputDate)) {
       this.setState({
         selectedDay: DateInput.getDate(inputDate, FORMATS.DATE_OBJECT, dateFormat),
       }, () => {
@@ -197,6 +221,19 @@ export default class DateInput extends React.Component {
     }
   };
 
+  handleInputBlur = () => {
+    this.prettifyInputDate();
+
+    // We want to close the overlay on blur, unless it was caused by a click on the calendar
+    // overlay
+    if (!this.mouseClickedOnContainer) {
+      this.setState({
+        showOverlay: false,
+      });
+    }
+    this.mouseClickedOnContainer = false;
+  };
+
   /**
    * Handles dayPicker click
    * @param day {date}
@@ -208,7 +245,8 @@ export default class DateInput extends React.Component {
     const momentDate = moment.utc(day);
 
     let timeAdjustedDate = null;
-    const currentMomentDate = moment(value, moment.ISO_8601).utc();
+    const currentMomentDate = moment(value, moment.ISO_8601)
+      .utc();
     const currentHours = currentMomentDate.get('hour');
     const currentMinutes = currentMomentDate.get('minute');
 
@@ -231,6 +269,8 @@ export default class DateInput extends React.Component {
       this.props.onChange(DateInput.getDate(timeAdjustedDate, FORMATS.UTC, dateFormat));
       this.input.blur();
     });
+
+    this.props.onDayClick(day, modifiers);
   };
 
   /**
@@ -249,14 +289,44 @@ export default class DateInput extends React.Component {
     });
   };
 
+
   /**
    * Handles year-month picker (select boxes) change
    * @param date
    */
   handleYearMonthChange = (val) => {
+    const { value, dateFormat } = this.props;
+    const momentDate = value ? moment.utc(value, moment.ISO_8601) : moment.utc();
+
+    momentDate.year(val.getFullYear())
+      .month(val.getMonth());
+
     this.setState({
+      inputDate: DateInput.getDate(momentDate, FORMATS.PRETTY_DATE, dateFormat),
+      selectedDay: DateInput.getDate(momentDate, FORMATS.DATE_OBJECT, dateFormat),
       dayPickerVisibleMonth: val,
+    }, () => {
+      this.props.onChange(DateInput.getDate(momentDate, FORMATS.UTC, dateFormat));
     });
+  };
+
+  /**
+   * Handles a click on the overlay
+   * @param e
+   */
+  handleOnOverlayMouseDown = (e) => {
+    if (this.calendarContainer.contains(e.target)) {
+      this.mouseClickedOnContainer = true;
+    }
+  };
+
+  /**
+   * Clears input value
+   */
+  handleClearClick = () => {
+    const { onChange } = this.props;
+    if (!onChange) throw new TypeError('react-datetime: onChange callback is not set');
+    this.props.onChange('');
   };
 
   /**
@@ -278,6 +348,14 @@ export default class DateInput extends React.Component {
     return pattern.test(date.trim());
   };
 
+  prettifyInputDate = () => {
+    const { value, dateFormat } = this.props;
+    const momentDate = moment.utc(value, moment.ISO_8601);
+    this.setState({
+      inputDate: DateInput.getDate(momentDate, FORMATS.PRETTY_DATE, dateFormat),
+    });
+  };
+
   /**
    * Renders select boxes above the calendar
    * @param date
@@ -291,18 +369,32 @@ export default class DateInput extends React.Component {
     />
   );
 
+  renderClearValueButton = () => (
+    <button
+      type="button"
+      className={this.props.disabled ?
+        `${classPrefix}-clear-value disabled` : `${classPrefix}-clear-value`}
+      onClick={this.handleClearClick}
+      disabled={this.props.disabled}
+    >
+      <span>x</span>
+    </button>
+  );
+
   render() {
-    const classPrefix = 'oc-datetime';
     /* eslint-disable no-unused-vars */
     const {
+      className,
       locale,
       time,
       value,
       inputProps,
       inputRef,
       disabled,
+      selectedDays,
       showWeekNumbers,
       minutesInterval,
+      showClearValue,
       ...otherProps
     } = this.props;
     const momentDate = moment.utc(value, moment.ISO_8601);
@@ -310,6 +402,8 @@ export default class DateInput extends React.Component {
       hour: momentDate.hour(),
       minute: momentDate.minute(),
     };
+    const month = this.state.dayPickerVisibleMonth
+      || ((typeof this.state.selectedDay === 'string') ? undefined : this.state.selectedDay);
 
     return (
       <TetherComponent
@@ -323,9 +417,9 @@ export default class DateInput extends React.Component {
             to: 'window',
             attachment: 'together',
           }]}
-        className={`${classPrefix}`}
+        className={`${classPrefix} ${className}`}
       >
-        <FormGroup>
+        <FormGroup className={`${classPrefix}-input-container`}>
           <FormControl
             type="text"
             inputRef={(el) => {
@@ -334,40 +428,50 @@ export default class DateInput extends React.Component {
             }}
             value={this.state.inputDate}
             disabled={disabled}
+            autoComplete="off"
             {...inputProps}
             onChange={this.handleInputChange}
             onFocus={this.handleInputFocus}
+            onBlur={this.handleInputBlur}
           />
+          {showClearValue && value && this.renderClearValueButton()}
         </FormGroup>
-        {this.state.showOverlay &&
+
+        {this.state.showOverlay
+        && (
         <div
           role="presentation"
           className={`${classPrefix}-calendar`}
           ref={(el) => {
             this.calendarContainer = el;
           }}
+          onMouseDown={this.handleOnOverlayMouseDown}
         >
           <DayPicker
+            {...otherProps}
             ref={(el) => {
               this.dayPicker = el;
             }}
-            onDayClick={this.handleDayClick}
-            selectedDays={this.isSameDay}
+            selectedDays={selectedDays || this.isSameDay}
             localeUtils={this.localeUtils}
-            month={this.state.dayPickerVisibleMonth}
+            month={month}
             showWeekNumbers={showWeekNumbers}
             firstDayOfWeek={this.getFirstDayOfWeek()}
             locale={locale}
             captionElement={this.renderCaptionElement}
-            {...otherProps}
+            navbarElement={Navbar}
+            onDayClick={this.handleDayClick}
           />
-          {time &&
+          {time
+          && (
           <TimePicker
             onChange={this.handleTimePickerChange}
             time={timeObj}
             minutesInterval={minutesInterval}
-          />}
+          />
+          )}
         </div>
+        )
         }
       </TetherComponent>
     );